Fixed setting up Elixir during CI runs
CI runs would bail out in the 'Set up Elixir' step, printing

  Error: Tried to map a target OS from env. variable 'ImageOS' (got
  ubuntu22), but failed. If you're using a self-hosted runner, you should
  set 'env': 'ImageOS': ... to one of the following: ['ubuntu16',
  'ubuntu18', 'ubuntu20', 'win16', 'win19']

It appears that this is because 'ubuntu-latest' corresponds to
'ubuntu22' but the setup-beam action doesn't know that. Looking at other
projects showed that they use 'runs-on' with a specific image version,
let's see if that works better.

The ubuntu-20.04 image has the widest range of supported OTP version as
per

  https://github.com/erlef/setup-beam#compatibility-between-operating-system-and-erlangotp

so let's use that.

Let's also not hardcode a setup-erlef hash but just use @v1.